release 0.3.0b

+++ data/db/src/main/java/leap/db/command/CreateForeignKey.java
@@ -19,7 +19,7 @@ import leap.db.DbCommand;
 
 public interface CreateForeignKey extends DbCommand {
 
-	default int getSortOrder() {
+	default float getSortOrder() {
 		return 3;
 	}
